requirements.txt
asgiref==3.7.2
Babel==2.14.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.8
django-allauth==0.59.0
django-bootstrap5==23.3
django-cloudinary-storage==0.3.0
django-phonenumber-field==7.2.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.26
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
